code/backend/requirements.txt
php
Copy
Edit
Flask
psycopg2-binary